[ADD][send_debo_fields] - made creation of debo fields separate from sending them
