+++ debo_integration_fields/models/__init__.py
@@ -5,3 +5,4 @@ from . import product_template
 from . import res_company
 from . import res_partner
 from . import res_config
+from . import res_users
