+++ send_debo_fields/models/__init__.py
@@ -1,0 +1,3 @@
+from . import res_partner
+from . import product_product
+from . import res_config
